+++ client/src/core/client/admin/routes/Reports/SingleReportRoute.css
@@ -42,57 +42,8 @@
   margin-left: auto;
 }
 
-.shareButton {
-  margin-left: var(--spacing-2);
-  background-color: var(--palette-background-body);
-}
-
 .reportDate {
   margin-left: auto;
-}
-
-.reportHistory {
-  width: 30%;
-  padding: var(--spacing-6) var(--spacing-4);
-}
-
-.reportHistoryHeader {
-  text-transform: uppercase;
-  font-size: var(--font-size-3);
-  font-weight: var(--font-weight-secondary-bold);
-  font-family: var(--font-family-secondary);
-  padding-bottom: var(--spacing-3);
-}
-
-.reportHistoryText {
-  font-size: var(--font-size-2);
-}
-
-.reportHistoryCreatedAt {
-  clear: both;
-  font-size: var(--font-size-1);
-  color: var(--palette-grey-500);
-}
-
-.reportHistoryNoteBody {
-  padding: var(--spacing-1);
-  background-color: var(--palette-grey-100);
-  margin-top: var(--spacing-1);
-  font-size: var(--font-size-2);
-}
-
-.deleteReportNoteButton {
-  float: right;
-  margin-top: var(--spacing-1);
-}
-
-.addNoteTextarea {
-  width: 100%;
-}
-
-.addNoteTextarea::placeholder {
-  color: var(--palette-grey-300);
-  font-size: var(--font-size-2);
 }
 
 .commentMain {
